requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic

# Google Gemini API
google-genai

# WebSocket Support
websockets

# HTTP Client (for external API calls if needed)
httpx

# Environment Variables Management
python-dotenv

# Logging (Optional - stdlib works, but this is better)
structlog
